Add support for key attestations in JWT proofs
