+++ src/main/kotlin/eu/europa/ec/eudi/openid4vci/internal/Jwt.kt
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2023 European Commission
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package eu.europa.ec.eudi.openid4vci.internal
+
+import com.nimbusds.jose.JWSAlgorithm
+import com.nimbusds.jose.JWSObject
+import com.nimbusds.jose.crypto.MACSigner
+import com.nimbusds.jwt.SignedJWT
+
+internal fun SignedJWT.ensureSignedNotMAC() {
+    check(state == JWSObject.State.SIGNED || state == JWSObject.State.VERIFIED) {
+        "Provided JWT is not signed"
+    }
+    val alg = requireNotNull(header.algorithm) { "Invalid JWT misses header alg" }
+    requireIsNotMAC(alg)
+}
+
+internal fun requireIsNotMAC(alg: JWSAlgorithm) =
+    require(!alg.isMACSigning()) { "MAC signing algorithm not allowed" }
+
+private fun JWSAlgorithm.isMACSigning(): Boolean = this in MACSigner.SUPPORTED_ALGORITHMS
